+++ test_problems/cathermo/HMW_graph_GvT/HMW_graph_GvT.cpp
@@ -25,7 +25,7 @@ int main(int argc, char** argv)
 {
 
     int retn = 0;
-    int i;
+    size_t i;
     int extraCols = 1;
 
     try {
@@ -46,7 +46,7 @@ int main(int argc, char** argv)
         string id = "NaCl(S)";
         Cantera::ThermoPhase* solid = Cantera::newPhase(nacl_s, id);
 
-        int nsp = HMW->nSpecies();
+        size_t nsp = HMW->nSpecies();
         double acMol[100];
         double act[100];
         double mf[100];
@@ -65,8 +65,8 @@ int main(int argc, char** argv)
 
         HMW->setState_TP(298.15, 1.01325E5);
 
-        int i1 = HMW->speciesIndex("Na+");
-        int i2 = HMW->speciesIndex("Cl-");
+        size_t i1 = HMW->speciesIndex("Na+");
+        size_t i2 = HMW->speciesIndex("Cl-");
         //int i3 = HMW->speciesIndex("H2O(L)");
         for (i = 1; i < nsp; i++) {
             moll[i] = 0.0;
@@ -167,7 +167,7 @@ int main(int argc, char** argv)
             mu_NaCl = mu[0] * 1.0E-6;
 
             HMW->getChemPotentials(mu);
-            for (int k = 0; k < nsp; k++) {
+            for (size_t k = 0; k < nsp; k++) {
                 mu[k] *= 1.0E-6;
             }
             mu_Naplus  = mu[i1];
@@ -184,16 +184,14 @@ int main(int argc, char** argv)
             double molecWater = HMW->molecularWeight(0);
             double Mo = molecWater / 1000.;
             double Gibbs_kgWater = molarGibbs / (Xmol[0] * Mo);
-
-
             double Aphi = HMW->A_Debye_TP() / 3.0;
 
-            for (int k = 0; k < nsp; k++) {
+            for (size_t k = 0; k < nsp; k++) {
                 mu0_RT[k] *= RT * 1.0E-6;
             }
 
             molarGibbs0 = 0.0;
-            for (int k = 0; k < nsp; k++) {
+            for (size_t k = 0; k < nsp; k++) {
                 molarGibbs0 += Xmol[k] * mu0_RT[k];
             }
             double Gibbs0_kgWater = molarGibbs0 / (Xmol[0] * Mo);
